aicloudops/chatbot/requirements.txt
pyautogen==0.2.19
docker==7.0.0
pika==1.3.2
python-dotenv==1.0.1
typing_extensions==4.10.0
kubernetes==29.0.0
pytest==8.1.1

